Refactor: move schdule function to utility/ranking
