+++ package.json
@@ -20,6 +20,7 @@
     "express-asyncify": "2.1.2",
     "mongodb": "6.3.0",
     "mongoose": "8.0.3",
+    "node-schedule": "2.1.1",
     "passport": "0.7.0",
     "passport-kakao": "1.0.1",
     "pg": "8.11.3",
